Fix: backend is now gcc standard Compliant

+++ backend/src/algo/metrics.c
@@ -48,20 +48,20 @@ int em_ch_stat(const double *class_table, int32_t samples, int32_t classes, int3
     if (!class_table || samples <= 0 || classes <= 0 || k <= 1) return -1;
 
     int i, j;
-    double *totsum = calloc(classes, sizeof(double));
-    double *totav  = calloc(classes, sizeof(double));
-    double *sst    = calloc(classes, sizeof(double));
-    double **clsum = calloc(k, sizeof(double*));
-    double *clsam  = calloc(k, sizeof(double));
-    double **clav  = calloc(k, sizeof(double*));
-    double **sse   = calloc(k, sizeof(double*));
+    double *totsum = calloc((size_t)classes, sizeof(double));
+    double *totav  = calloc((size_t)classes, sizeof(double));
+    double *sst    = calloc((size_t)classes, sizeof(double));
+    double **clsum = calloc((size_t)k, sizeof(double*));
+    double *clsam  = calloc((size_t)k, sizeof(double));
+    double **clav  = calloc((size_t)k, sizeof(double*));
+    double **sse   = calloc((size_t)k, sizeof(double*));
 
     if (!totsum || !totav || !sst || !clsum || !clsam || !clav || !sse) return -1;
 
     for (i = 0; i < k; i++) {
-        clsum[i] = calloc(classes, sizeof(double));
-        clav[i] = calloc(classes, sizeof(double));
-        sse[i] = calloc(classes, sizeof(double));
+        clsum[i] = calloc((size_t)classes, sizeof(double));
+        clav[i] = calloc((size_t)classes, sizeof(double));
+        sse[i] = calloc((size_t)classes, sizeof(double));
         if (!clsum[i] || !clav[i] || !sse[i]) return -1;
     }
 
@@ -125,11 +125,11 @@ int em_ch_stat(const double *class_table, int32_t samples, int32_t classes, int3
     if (perms_n > 0) {
         double perm_sum = 0.0;
         int perm_better = 0;
-        double *perm_data = malloc(samples * (classes + 1) * sizeof(double));
+        double *perm_data = malloc((size_t)samples * (size_t)(classes + 1) * sizeof(double));
         if (!perm_data) goto cleanup;
 
-        memcpy(perm_data, class_table, samples * (classes + 1) * sizeof(double));
-        srand(seed);
+        memcpy(perm_data, class_table, (size_t)samples * (size_t)(classes + 1) * sizeof(double));
+        srand((unsigned int)seed);
 
         for (int p = 0; p < perms_n; p++) {
             for (j = 0; j < classes; j++) {
